requirements.txt
asgiref==3.6.0
Django==4.2
django-ranged-response==0.2.0
django-simple-captcha==0.5.17
Pillow==9.5.0
psycopg2==2.9.6
sqlparse==0.4.3
tzdata==2023.3
Unipath==1.1